+++ client/package.json
@@ -12,7 +12,7 @@
     "react-dom": "^16.13.1",
     "react-icons": "^3.11.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "react-scripts": "^3.4.4",
     "react-spring": "^8.0.27",
     "resize-observer-polyfill": "^1.5.1",
     "styled-components": "^5.2.0"
